+++ test/vector.test.js
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import { Vector } from '../src';
+import { DimensionalityMismatchError, Line, Vector } from '../src';
 import { asDiagram } from './_as-diagram';
 
 describe('vector', () => {
@@ -110,22 +110,41 @@ describe('vector', () => {
     expectCall(x).augment(new Vector([5])).to.vector.equal([3, 4, 5]);
   });
 
-  it('should log', () => {
-    expect(x.log()).to.vector.equal([1.0986122886681098, 1.386294361119890]);
+  asDiagram('Vector.log').it(expectCall => {
+    expectCall(x).log().to.vector.equal([1.0986122886681098, 1.386294361119890]);
+    expectCall(x).log(3).to.vector.equal([1, 1.261859507142914]);
   });
 
-  it('should support removal of head positions', () => {
-    expect(x.chomp(1)).to.vector.equal([4]);
+  asDiagram('Vector.cross').it(expectCall => {
+    expectCall(new Vector([1, 2, 3]))
+      .cross(new Vector([4, 5, 6]))
+      .to.vector.equal([-3, 6, -3]);
   });
 
-  it('should sum', () => {
-    expect(x.sum()).to.equal(7);
+  asDiagram('Vector.max').it(expectCall => {
+    expectCall(x).max().to.equal(4);
   });
 
-  it('should support addition of elements on the right side', () => {
-    // expect(x.augment([4, 5])).to.vector.equal([1, 2, 3, 4, 5]);
+  asDiagram('Vector.maxIndex').it(expectCall => {
+    expectCall(x).max().to.equal(4);
   });
 
+  asDiagram('Vector.indexOf').it(expectCall => {
+    expectCall(x).indexOf(0).to.equal(-1);
+    expectCall(x).indexOf(3).to.equal(1);
+  });
+
+  asDiagram('Vector.toDiagonalMatrix').it(expectCall => {
+    expectCall(x).toDiagonalMatrix().to.matrix.equal([[3, 0], [0, 4]]);
+  });
+
+  asDiagram('Vector.round').it(expectCall => {
+    expectCall(new Vector([1.5, 2.3])).round().to.vector.equal([2, 2]);
+  });
+
+  asDiagram('Vector.transpose').it(expectCall => {
+    expectCall(x).transpose().to.matrix.equal([[3], [4]]);
+  });
   it('show allow for scalar addition', () => {
     const a = new Vector([2, 3, 4]);
     const b = a.add(1);
@@ -136,5 +155,33 @@ describe('vector', () => {
     const a = new Vector([2, 3, 4]);
     const b = a.add(new Vector([2, 4, 8]));
     expect(b).to.vector.equal([undefined, 7, 12]);
+  });
+
+  describe('rotation', () => {
+    const start2d = new Vector([2, 3]);
+    const pivot2d = new Vector([1, 1]);
+    const start3d = new Vector([1, 2, 3]);
+    const pivot3d = new Line.create([0, 0, 0], [1, 1, 1]);
+
+    it('throws if cannot pivot', () => {
+      expect(() => new Vector([]).rotate(Math.PI, pivot2d))
+        .to.throw(DimensionalityMismatchError, /cannot be rotated/);
+      expect(() => new Vector([1]).rotate(Math.PI, pivot2d))
+        .to.throw(DimensionalityMismatchError, /cannot be rotated/);
+      expect(() => new Vector([1, 2, 3, 4]).rotate(Math.PI, pivot2d))
+        .to.throw(DimensionalityMismatchError, /cannot be rotated/);
+    });
+
+    it('pivots in 2d', () => {
+      expect(start2d.rotate(Math.PI, pivot2d)).to.vector.equal([-2.5, -1]);
+    });
+
+    it('pivots in 3d', () => {
+      expect(start3d.rotate(Math.PI, pivot3d)).to.vector.equal([
+        3,
+        2,
+        1
+      ]);
+    });
   });
 });